feat: extract loading screen to separate module
- Create loading_screen.h and loading_screen.cpp modules
- Move loading screen logic from main.cpp to the new module
- Update main.cpp to use the new loading screen functions
- Add InitLoadingScreen, UpdateLoadingScreen, RenderLoadingScreen functions
- Actual asset loading implementation with stage-based loading sequence

+++ include/screen.h
@@ -52,13 +52,13 @@ using GameState = struct
     ScreenState previousScreen;/**< Screen sebelum OPTIONS - buat return button */
     bool showFPS;              /**< Tampilkan FPS counter di HUD */
     
-    /*==============================================================================
+     /*==============================================================================
      * Loading State Variables
      *==============================================================================*/
     /**
      * @brief Current loading progress percentage (0-100)
      */
-    int loadingProgress;
+    float loadingProgress;
     /**
      * @brief Current loading status text to display
      */
